Added some of Jeromes suggestions, moved Ontology stuff to ontologies.py
