+++ tests/unit/test_imports.py
@@ -159,7 +159,8 @@ class ImportGraphLayerChecker(object):
         'datamodel': ['ga4gh/datamodel/reads.py',
                       'ga4gh/datamodel/references.py',
                       'ga4gh/datamodel/variants.py',
-                      'ga4gh/datamodel/datasets.py'],
+                      'ga4gh/datamodel/datasets.py',
+                      'ga4gh/datamodel/ontologies.py'],
         'libraries': ['ga4gh/converters.py',
                       'ga4gh/configtest.py'],
         'protocol': ['ga4gh/protocol.py',
